moves common types to types file (and out of index.ts)
