+++ src/components/combo_box/matching_options.test.ts
@@ -1,4 +1,4 @@
-import { EuiComboBoxOptionOption } from './index';
+import { EuiComboBoxOptionOption } from './types';
 import {
   flattenOptionGroups,
   getMatchingOptions,
