+++ src/components/combo_box/types.ts
@@ -1,4 +1,16 @@
-import { EuiComboBoxOptionOption } from '.';
+import { ButtonHTMLAttributes } from 'react';
+import { CommonProps } from '../common';
+
+// note similarity to `Option` in `components/selectable/types.tsx`
+export type EuiComboBoxOptionOption<
+  T = string | number | string[] | undefined
+> = CommonProps &
+  Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'value'> & {
+    isGroupLabelOption?: boolean;
+    label: string;
+    options?: Array<EuiComboBoxOptionOption<T>>;
+    value?: T;
+  };
 
 export type UpdatePositionHandler = (
   listElement?: RefInstance<HTMLDivElement>
@@ -11,3 +23,9 @@ export type RefCallback<T> = {
 }['bivarianceHack'];
 
 export type RefInstance<T> = T | null;
+
+export type EuiComboBoxOptionsListPosition = 'top' | 'bottom';
+
+export interface EuiComboBoxSingleSelectionShape {
+  asPlainText?: boolean;
+}
